refactor(sessions): separate SQLite accessor owners
