+++ scripts/lib/session-accessor-debt-baseline.json
@@ -9,7 +9,6 @@
     "src/commands/doctor-heartbeat-session-target.ts": 2,
     "src/commands/doctor-state-integrity.ts": 2,
     "src/commands/doctor/shared/codex-route-warnings.ts": 3,
-    "src/config/sessions/session-accessor.sqlite.ts": 2,
     "src/config/sessions/session-accessor.ts": 12,
     "src/config/sessions/store-load.ts": 3,
     "src/config/sessions/store.ts": 10,
@@ -29,7 +28,7 @@
   },
   "sessionCompactManualTrim": {},
   "sessionLifecycleCleanup": {
-    "src/config/sessions/session-accessor.sqlite.ts": 2,
+    "src/config/sessions/session-accessor.sqlite-projection.ts": 2,
     "src/config/sessions/store-maintenance-operations.ts": 2,
     "src/config/sessions/store.ts": 2
   },
@@ -38,7 +37,7 @@
     "src/agents/embedded-agent-runner/thinking-replay-repair.ts": 2,
     "src/agents/embedded-agent-runner/tool-result-truncation.ts": 3,
     "src/agents/embedded-agent-runner/transcript-rewrite.ts": 3,
-    "src/config/sessions/session-accessor.sqlite.ts": 3,
+    "src/config/sessions/session-accessor.sqlite-events.ts": 3,
     "src/config/sessions/session-accessor.ts": 2
   }
 }
