+++ src/config/sessions/session-accessor.sqlite-entry-store.ts
@@ -1,0 +1,468 @@
+import { uniqueStrings } from "@openclaw/normalization-core/string-normalization";
+import type { Selectable } from "kysely";
+import {
+  executeSqliteQuerySync,
+  executeSqliteQueryTakeFirstSync,
+} from "../../infra/kysely-sync.js";
+import type { DB as OpenClawAgentKyselyDatabase } from "../../state/openclaw-agent-db.generated.js";
+import type { OpenClawAgentDatabase } from "../../state/openclaw-agent-db.js";
+import { normalizeSqliteNumber } from "./session-accessor.sqlite-normalize.js";
+import { resolveSessionEntryProvenanceRow } from "./session-accessor.sqlite-provenance.js";
+import {
+  cloneSessionEntry,
+  getSessionKysely,
+  normalizeSqliteSessionKey,
+} from "./session-accessor.sqlite-scope.js";
+import {
+  bindSqliteSessionRoot,
+  normalizeSqliteSessionEntryTimestamp,
+} from "./session-accessor.sqlite-session-row.js";
+import {
+  normalizeSqliteStatus,
+  parseSqliteSessionEntryJson as parseSessionEntryRow,
+} from "./session-accessor.sqlite-status.js";
+import {
+  readTranscriptMutationStateInTransaction,
+  writeSessionRoute,
+} from "./session-accessor.sqlite-transcript-state.js";
+import {
+  foldedSessionKeyAliasCandidates,
+  normalizeStoreSessionKey,
+  resolveSessionEntryCandidates,
+} from "./store-entry.js";
+import type { SessionEntry } from "./types.js";
+
+// Canonical owner for session_entries row selection, alias snapshots, and writes.
+
+type SessionEntryRow = Selectable<OpenClawAgentKyselyDatabase["session_entries"]>;
+export type ResolvedSessionEntryRow = {
+  entry: SessionEntry;
+  legacyKeys: string[];
+  row: SessionEntryRow;
+};
+export type SqliteSessionEntrySelectionSnapshot = {
+  selected: ResolvedSessionEntryRow | undefined;
+  selectedRows: Array<{ entry: SessionEntry; sessionKey: string }>;
+};
+export type SqliteLifecycleTargetSnapshot = {
+  primary: { entry: SessionEntry; key: string } | undefined;
+  rows: Array<{ entry: SessionEntry; sessionKey: string }>;
+};
+
+class SqliteSessionMutationConflictError extends Error {
+  constructor(operationLabel: string) {
+    super(`SQLite session state changed while preparing ${operationLabel}`);
+    this.name = "SqliteSessionMutationConflictError";
+  }
+}
+
+export function readSqliteSessionIdentitySnapshot(
+  database: OpenClawAgentDatabase,
+  sessionKeys: Iterable<string>,
+): Map<string, SessionEntry> {
+  const snapshot = new Map<string, SessionEntry>();
+  for (const sessionKey of uniqueStrings([...sessionKeys].map((key) => key.trim()))) {
+    const row = readExactSessionEntryRow(database, sessionKey);
+    if (row) {
+      snapshot.set(sessionKey, cloneSessionEntry(row.entry));
+    }
+  }
+  return snapshot;
+}
+
+export function createSqliteSessionIdentitySnapshot(
+  rows: readonly { entry: SessionEntry; sessionKey: string }[],
+): Map<string, SessionEntry> {
+  return new Map(rows.map((row) => [row.sessionKey, cloneSessionEntry(row.entry)]));
+}
+
+export function readSessionEntryRow(
+  database: OpenClawAgentDatabase,
+  sessionKey: string,
+): ResolvedSessionEntryRow | undefined {
+  const db = getSessionKysely(database.db);
+  const lookupKeys = collectSessionEntryLookupKeys(database, sessionKey);
+  if (lookupKeys.length === 0) {
+    return undefined;
+  }
+  const rows = executeSqliteQuerySync(
+    database.db,
+    db
+      .selectFrom("session_entries")
+      .selectAll()
+      .where("session_key", "in", lookupKeys)
+      .orderBy("session_key", "asc"),
+  ).rows;
+  const entries = new Map<string, ResolvedSessionEntryRow>();
+  for (const row of rows) {
+    const entry = parseSessionEntryRow(row);
+    if (!entry) {
+      continue;
+    }
+    entries.set(row.session_key, { entry, legacyKeys: [], row });
+  }
+  const resolved = resolveSessionEntryCandidates({
+    entries: [...entries].map(([candidateKey, value]) => ({
+      entry: value.entry,
+      sessionKey: candidateKey,
+    })),
+    sessionKey,
+  });
+  if (!resolved.existing) {
+    return undefined;
+  }
+  const selected = entries.get(resolved.existing.sessionKey);
+  return selected ? { ...selected, legacyKeys: resolved.legacyKeys } : undefined;
+}
+
+// Async updaters prepare against this complete selection. Capturing alias rows
+// prevents the commit phase from deleting a concurrently changed legacy key.
+export function readSqliteSessionEntrySelectionSnapshot(
+  database: OpenClawAgentDatabase,
+  sessionKey: string,
+  exact: boolean,
+): SqliteSessionEntrySelectionSnapshot {
+  const selected = exact
+    ? readExactSessionEntryRow(database, sessionKey)
+    : readSessionEntryRow(database, sessionKey);
+  const selectedKeys = collectSessionEntryLookupKeys(database, sessionKey).toSorted();
+  return {
+    selected,
+    selectedRows: selectedKeys.flatMap((candidateKey) => {
+      const row = readExactSessionEntryRow(database, candidateKey);
+      return row ? [{ entry: cloneSessionEntry(row.entry), sessionKey: candidateKey }] : [];
+    }),
+  };
+}
+
+export function assertSqliteSessionEntrySelectionUnchanged(
+  expected: SqliteSessionEntrySelectionSnapshot,
+  current: SqliteSessionEntrySelectionSnapshot,
+  operationLabel: string,
+): void {
+  const selectedMatches =
+    expected.selected?.row.session_key === current.selected?.row.session_key &&
+    sqliteSessionEntriesEqual(expected.selected?.entry, current.selected?.entry);
+  if (
+    !selectedMatches ||
+    !sqliteSessionSnapshotRowsEqual(expected.selectedRows, current.selectedRows)
+  ) {
+    throw new SqliteSessionMutationConflictError(operationLabel);
+  }
+}
+
+export function collectSessionEntryLookupKeys(
+  database: OpenClawAgentDatabase,
+  sessionKey: string,
+): string[] {
+  const trimmedKey = sessionKey.trim();
+  if (!trimmedKey) {
+    return [];
+  }
+  const normalizedKey = normalizeStoreSessionKey(trimmedKey);
+  const lookupKeys = new Set([
+    trimmedKey,
+    normalizedKey,
+    ...foldedSessionKeyAliasCandidates(normalizedKey),
+  ]);
+  const db = getSessionKysely(database.db);
+  const rows = executeSqliteQuerySync(
+    database.db,
+    db.selectFrom("session_entries").select("session_key").orderBy("session_key", "asc"),
+  ).rows;
+  for (const row of rows) {
+    if (normalizeStoreSessionKey(row.session_key) === normalizedKey) {
+      lookupKeys.add(row.session_key);
+    }
+  }
+  return [...lookupKeys].filter(Boolean);
+}
+
+export function readExactSessionEntryRow(
+  database: OpenClawAgentDatabase,
+  sessionKey: string,
+): ResolvedSessionEntryRow | undefined {
+  const db = getSessionKysely(database.db);
+  const row = executeSqliteQueryTakeFirstSync(
+    database.db,
+    db.selectFrom("session_entries").selectAll().where("session_key", "=", sessionKey),
+  );
+  if (!row) {
+    return undefined;
+  }
+  const entry = parseSessionEntryRow(row);
+  return entry ? { entry, legacyKeys: [], row } : undefined;
+}
+
+export function readSqliteSessionEntryStore(
+  database: OpenClawAgentDatabase,
+): Record<string, SessionEntry> {
+  const db = getSessionKysely(database.db);
+  const rows = executeSqliteQuerySync(
+    database.db,
+    db.selectFrom("session_entries").select(["session_key", "entry_json"]).orderBy("session_key"),
+  ).rows;
+  const store: Record<string, SessionEntry> = {};
+  for (const row of rows) {
+    const entry = parseSessionEntryRow(row);
+    if (entry) {
+      store[row.session_key] = entry;
+    }
+  }
+  return store;
+}
+
+export function readSqliteSessionEntryCount(database: OpenClawAgentDatabase): number {
+  const db = getSessionKysely(database.db);
+  const row = executeSqliteQueryTakeFirstSync(
+    database.db,
+    db.selectFrom("session_entries").select((eb) => eb.fn.countAll<number>().as("entry_count")),
+  );
+  const count = row?.entry_count;
+  return count === undefined || count === null ? 0 : normalizeSqliteNumber(count);
+}
+
+export function resolveSqliteLifecyclePrimaryEntry(
+  database: OpenClawAgentDatabase,
+  target: { canonicalKey: string; storeKeys: string[] },
+): { key: string; entry: SessionEntry } | undefined {
+  let freshest: { key: string; entry: SessionEntry } | undefined;
+  for (const key of target.storeKeys) {
+    const row = readExactSessionEntryRow(database, key.trim());
+    if (!row) {
+      continue;
+    }
+    if (!freshest || (row.entry.updatedAt ?? 0) > (freshest.entry.updatedAt ?? 0)) {
+      freshest = { key, entry: row.entry };
+    }
+  }
+  return freshest ?? undefined;
+}
+
+export function readSqliteLifecycleTargetSnapshot(
+  database: OpenClawAgentDatabase,
+  target: { canonicalKey: string; storeKeys: string[] },
+): SqliteLifecycleTargetSnapshot {
+  const normalized = normalizeSqliteLifecycleTarget(target);
+  return {
+    primary: resolveSqliteLifecyclePrimaryEntry(database, normalized),
+    rows: normalized.storeKeys.flatMap((sessionKey) => {
+      const row = readExactSessionEntryRow(database, sessionKey);
+      return row ? [{ entry: cloneSessionEntry(row.entry), sessionKey }] : [];
+    }),
+  };
+}
+
+export function assertSqliteLifecycleTargetSnapshotUnchanged(
+  expected: SqliteLifecycleTargetSnapshot,
+  current: SqliteLifecycleTargetSnapshot,
+  operationLabel: string,
+): void {
+  const primaryMatches =
+    expected.primary?.key === current.primary?.key &&
+    sqliteSessionEntriesEqual(expected.primary?.entry, current.primary?.entry);
+  if (!primaryMatches || !sqliteSessionSnapshotRowsEqual(expected.rows, current.rows)) {
+    throw new SqliteSessionMutationConflictError(operationLabel);
+  }
+}
+
+export function normalizeSqliteLifecycleTarget(target: {
+  canonicalKey: string;
+  storeKeys: string[];
+}): {
+  canonicalKey: string;
+  storeKeys: string[];
+} {
+  const canonicalKey = normalizeSqliteSessionKey(target.canonicalKey);
+  return {
+    canonicalKey,
+    storeKeys: uniqueStrings([canonicalKey, ...target.storeKeys.map(normalizeSqliteSessionKey)]),
+  };
+}
+
+export function deleteSqliteSessionEntryRows(
+  database: OpenClawAgentDatabase,
+  sessionKey: string,
+): void {
+  const db = getSessionKysely(database.db);
+  executeSqliteQuerySync(
+    database.db,
+    db.deleteFrom("session_routes").where("session_key", "=", sessionKey),
+  );
+  executeSqliteQuerySync(
+    database.db,
+    db.deleteFrom("session_entries").where("session_key", "=", sessionKey),
+  );
+}
+
+export function deleteSqliteLifecycleTargetRows(
+  database: OpenClawAgentDatabase,
+  target: { canonicalKey: string; storeKeys: string[] },
+): void {
+  for (const sessionKey of uniqueStrings([target.canonicalKey, ...target.storeKeys])) {
+    const trimmed = sessionKey.trim();
+    if (trimmed) {
+      deleteSqliteSessionEntryRows(database, trimmed);
+    }
+  }
+}
+
+export function sqliteSessionEntriesEqual(
+  left: SessionEntry | undefined,
+  right: SessionEntry | undefined,
+): boolean {
+  if (!left || !right) {
+    return left === right;
+  }
+  return JSON.stringify(left) === JSON.stringify(right);
+}
+
+export function sqliteSessionSnapshotRowsEqual(
+  left: Array<{ entry: SessionEntry; sessionKey: string }>,
+  right: Array<{ entry: SessionEntry; sessionKey: string }>,
+): boolean {
+  return (
+    left.length === right.length &&
+    left.every(
+      (row, index) =>
+        row.sessionKey === right[index]?.sessionKey &&
+        sqliteSessionEntriesEqual(row.entry, right[index]?.entry),
+    )
+  );
+}
+
+function sqliteLifecycleTargetMatchesExpectedEntry(
+  database: OpenClawAgentDatabase,
+  target: { canonicalKey: string; storeKeys: string[] },
+  expectedEntry: SessionEntry | undefined,
+): boolean {
+  const current = resolveSqliteLifecyclePrimaryEntry(database, target)?.entry;
+  if (!current || !expectedEntry) {
+    return current === expectedEntry;
+  }
+  return sqliteSessionEntriesEqual(current, expectedEntry);
+}
+
+export function assertSqliteLifecycleTargetUnchanged(
+  database: OpenClawAgentDatabase,
+  target: { canonicalKey: string; storeKeys: string[] },
+  expectedEntry: SessionEntry | undefined,
+  operation: "deleted" | "reset",
+): void {
+  if (sqliteLifecycleTargetMatchesExpectedEntry(database, target, expectedEntry)) {
+    return;
+  }
+  throw new Error(`SQLite session entry changed before ${operation} lifecycle mutation`);
+}
+
+export function deleteLegacySessionEntryRows(
+  database: OpenClawAgentDatabase,
+  legacyKeys: string[],
+  sessionKey: string,
+): void {
+  if (legacyKeys.length === 0) {
+    return;
+  }
+  const db = getSessionKysely(database.db);
+  for (const legacyKey of legacyKeys) {
+    if (legacyKey === sessionKey) {
+      continue;
+    }
+    executeSqliteQuerySync(
+      database.db,
+      db.deleteFrom("session_routes").where("session_key", "=", legacyKey),
+    );
+    executeSqliteQuerySync(
+      database.db,
+      db.deleteFrom("session_entries").where("session_key", "=", legacyKey),
+    );
+  }
+}
+
+export function writeSessionEntry(
+  database: OpenClawAgentDatabase,
+  sessionKey: string,
+  entry: SessionEntry,
+): void {
+  const db = getSessionKysely(database.db);
+  const normalizedEntry = normalizeSqliteSessionEntryTimestamp(entry);
+  const updatedAt = normalizedEntry.updatedAt;
+  const previousEntry = readExactSessionEntryRow(database, sessionKey)?.entry;
+  // Registry writes snapshot the current transcript watermark so recovery can
+  // distinguish same-millisecond transcript writes before and after this row.
+  const transcriptObservedAt =
+    readTranscriptMutationStateInTransaction(database, normalizedEntry.sessionId).updatedAt ??
+    updatedAt;
+  const boundSessionRoot = bindSqliteSessionRoot({
+    entry: normalizedEntry,
+    sessionKey,
+    updatedAt,
+  });
+  const boundSessionRow = {
+    ...boundSessionRoot,
+    transcript_observed_at: transcriptObservedAt,
+  };
+  const sessionRow = resolveSessionEntryProvenanceRow({
+    boundSessionRow,
+    database,
+    entry: normalizedEntry,
+    previousEntry,
+  });
+  executeSqliteQuerySync(
+    database.db,
+    db
+      .insertInto("sessions")
+      .values(sessionRow)
+      .onConflict((conflict) =>
+        conflict.column("session_id").doUpdateSet({
+          session_key: sessionKey,
+          session_scope: sessionRow.session_scope,
+          transcript_observed_at: transcriptObservedAt,
+          session_entry_provenance: sessionRow.session_entry_provenance,
+          acp_owned: sessionRow.acp_owned,
+          plugin_owner_id: sessionRow.plugin_owner_id,
+          hook_external_content_source: sessionRow.hook_external_content_source,
+          updated_at: updatedAt,
+          started_at: sessionRow.started_at,
+          ended_at: sessionRow.ended_at,
+          status: sessionRow.status,
+          chat_type: sessionRow.chat_type,
+          channel: sessionRow.channel,
+          account_id: sessionRow.account_id,
+          model_provider: sessionRow.model_provider,
+          model: sessionRow.model,
+          agent_harness_id: sessionRow.agent_harness_id,
+          parent_session_key: sessionRow.parent_session_key,
+          spawned_by: sessionRow.spawned_by,
+          display_name: sessionRow.display_name,
+        }),
+      ),
+  );
+  writeSessionRoute(database, {
+    sessionId: sessionRow.session_id,
+    sessionKey,
+    updatedAt,
+  });
+  executeSqliteQuerySync(
+    database.db,
+    db
+      .insertInto("session_entries")
+      .values({
+        session_key: sessionKey,
+        session_id: normalizedEntry.sessionId,
+        entry_json: JSON.stringify(normalizedEntry),
+        updated_at: updatedAt,
+        status: normalizeSqliteStatus(normalizedEntry.status),
+      })
+      .onConflict((conflict) =>
+        conflict.column("session_key").doUpdateSet({
+          session_id: normalizedEntry.sessionId,
+          entry_json: JSON.stringify(normalizedEntry),
+          updated_at: updatedAt,
+          status: normalizeSqliteStatus(normalizedEntry.status),
+        }),
+      ),
+  );
+}
+
+/** Resolves the parent fork decision using SQLite transcript rows when totals are stale. */
