+++ src/config/sessions/session-accessor.sqlite-transcript-write.ts
@@ -1,0 +1,602 @@
+import { randomUUID } from "node:crypto";
+import { resolveTimestampMsToIsoString } from "@openclaw/normalization-core/number-coercion";
+import {
+  openOpenClawAgentDatabase,
+  runOpenClawAgentWriteTransaction,
+  type OpenClawAgentDatabase,
+} from "../../state/openclaw-agent-db.js";
+import type {
+  SessionTranscriptAccessScope,
+  SessionTranscriptTurnMessageAppend,
+  SessionTranscriptTurnWriteContext,
+  SessionTranscriptWriteScope,
+  TranscriptEvent,
+  TranscriptMessageAppendOptions,
+  TranscriptMessageAppendResult,
+} from "./session-accessor.sqlite-contract.js";
+import type { ResolvedSessionEntryRow } from "./session-accessor.sqlite-entry-store.js";
+import {
+  collectSessionEntryLookupKeys,
+  deleteLegacySessionEntryRows,
+  readSessionEntryRow,
+  readSqliteSessionIdentitySnapshot,
+  writeSessionEntry,
+} from "./session-accessor.sqlite-entry-store.js";
+import { emitCommittedSessionIdentityDiff } from "./session-accessor.sqlite-identity.js";
+import {
+  readSqliteTranscriptSnapshot,
+  readTranscriptEventJsonSetInTransaction,
+  type SqliteTranscriptSnapshotRow,
+} from "./session-accessor.sqlite-read.js";
+import {
+  cloneSessionEntry,
+  formatSqliteSessionMarkerForScope,
+  resolveSqliteScope,
+  resolveSqliteTranscriptScope,
+  runExclusiveSqliteSessionWrite,
+  toDatabaseOptions,
+  type ResolvedTranscriptScope,
+} from "./session-accessor.sqlite-scope.js";
+import {
+  advanceTranscriptMutationAtInTransaction,
+  touchTranscriptMutationInTransaction,
+} from "./session-accessor.sqlite-transcript-state.js";
+import {
+  appendTranscriptEventInTransaction,
+  ensureTranscriptHeader,
+  readActiveTranscriptAppendParentId,
+  readMessageIdempotencyKey,
+  readTranscriptMessageByEventId,
+  readTranscriptMessageByScopedIdempotencyKey,
+  redactTranscriptMessageForStorage,
+  replaceSqliteTranscriptEventsInTransaction,
+} from "./session-accessor.sqlite-transcript-store.js";
+import type {
+  SessionTranscriptTurnExpectedState,
+  SessionTranscriptTurnLifecyclePatch,
+} from "./session-transcript-turn-lifecycle.types.js";
+import {
+  buildExpectedTranscriptTurnSessionPatch,
+  sessionMatchesExpectedTranscriptTurn,
+} from "./session-transcript-turn-state.js";
+import type { SessionEntry } from "./types.js";
+import { mergeSessionEntry } from "./types.js";
+
+// Transcript write owner. Queue coordination surrounds synchronous SQLite commit sections.
+
+class SqliteTranscriptMutationConflictError extends Error {
+  constructor(sessionId: string) {
+    super(`SQLite transcript changed while preparing rewrite for ${sessionId}`);
+    this.name = "SqliteTranscriptMutationConflictError";
+  }
+}
+
+/** Internal doctor/migration import target for one legacy session row. */
+type SqliteSessionImportRowsParams = {
+  agentId?: string;
+  env?: NodeJS.ProcessEnv;
+  storePath?: string;
+  sessionKey: string;
+  entry: SessionEntry;
+  readTranscriptEvents?: (append: (event: TranscriptEvent) => void) => void;
+  transcriptMtimeMs?: number;
+};
+
+/** Summary of rows written by an internal doctor/migration import. */
+type SqliteSessionImportRowsResult = {
+  sessionId: string;
+  sessionKey: string;
+  transcriptEvents: number;
+};
+
+type SqliteExpectedSessionTranscriptTurnResult = {
+  appendedMessages: TranscriptMessageAppendResult<unknown>[];
+  rejectedReason?: "session-rebound";
+  sessionEntry: SessionEntry | undefined;
+  sessionFile: string;
+};
+
+type SqliteTranscriptWriteLockContext = {
+  appendMessage: <TMessage>(
+    options: TranscriptMessageAppendOptions<TMessage>,
+  ) => Promise<TranscriptMessageAppendResult<TMessage> | undefined>;
+  readEvents: () => Promise<TranscriptEvent[]>;
+  replaceEvents: (events: readonly TranscriptEvent[]) => Promise<void>;
+};
+
+type SqliteTranscriptSnapshotState =
+  | { kind: "current"; rows: SqliteTranscriptSnapshotRow[] }
+  | { kind: "stale" };
+
+export async function replaceSqliteTranscriptEvents(
+  scope: SessionTranscriptAccessScope,
+  events: TranscriptEvent[],
+): Promise<void> {
+  const resolved = resolveSqliteTranscriptScope(scope);
+  await runExclusiveSqliteSessionWrite(resolved, async () => {
+    runOpenClawAgentWriteTransaction((database) => {
+      replaceSqliteTranscriptEventsInTransaction(database, resolved, events);
+    }, toDatabaseOptions(resolved));
+  });
+}
+
+/** Fully replaces rows for one transcript synchronously for sync session runtimes. */
+export function replaceSqliteTranscriptEventsSync(
+  scope: SessionTranscriptAccessScope,
+  events: TranscriptEvent[],
+): boolean {
+  const resolved = resolveSqliteTranscriptScope(scope);
+  let replaced = false;
+  runOpenClawAgentWriteTransaction((database) => {
+    const fresh = readSessionEntryRow(database, resolved.sessionKey);
+    if (!fresh || fresh.entry.sessionId !== resolved.sessionId) {
+      return;
+    }
+    replaceSqliteTranscriptEventsInTransaction(database, resolved, events);
+    replaced = true;
+  }, toDatabaseOptions(resolved));
+  return replaced;
+}
+
+/** Imports one legacy session entry and its transcript rows for doctor migration. */
+export async function importSqliteSessionRows(
+  params: SqliteSessionImportRowsParams,
+): Promise<SqliteSessionImportRowsResult> {
+  const resolved = resolveSqliteScope({
+    ...(params.agentId ? { agentId: params.agentId } : {}),
+    ...(params.env ? { env: params.env } : {}),
+    sessionKey: params.sessionKey,
+    ...(params.storePath ? { storePath: params.storePath } : {}),
+  });
+  return await runExclusiveSqliteSessionWrite(resolved, async () => {
+    let transcriptEvents = 0;
+    runOpenClawAgentWriteTransaction((database) => {
+      const currentEntry = readSessionEntryRow(database, resolved.sessionKey)?.entry;
+      const preservedHarnessId =
+        params.entry.agentHarnessId === undefined &&
+        currentEntry?.sessionId === params.entry.sessionId &&
+        currentEntry.lifecycleRevision === params.entry.lifecycleRevision
+          ? currentEntry.agentHarnessId?.trim()
+          : undefined;
+      // Plugin doctor migrations can claim a legacy session before the full
+      // session import runs. Preserve that same-generation canonical owner.
+      const importedEntry = {
+        ...params.entry,
+        ...(preservedHarnessId ? { agentHarnessId: preservedHarnessId } : {}),
+        sessionFile: formatSqliteSessionMarkerForScope({
+          ...resolved,
+          sessionId: params.entry.sessionId,
+        }),
+      };
+      writeSessionEntry(database, resolved.sessionKey, importedEntry);
+      if (params.readTranscriptEvents) {
+        const transcriptScope = {
+          ...resolved,
+          sessionId: params.entry.sessionId,
+        };
+        const existingEventJson = readTranscriptEventJsonSetInTransaction(
+          database,
+          params.entry.sessionId,
+        );
+        params.readTranscriptEvents((event) => {
+          const eventJson = JSON.stringify(event);
+          if (existingEventJson.has(eventJson)) {
+            return;
+          }
+          if (
+            appendTranscriptEventInTransaction(database, transcriptScope, event, {
+              touchMutation: false,
+            })
+          ) {
+            existingEventJson.add(eventJson);
+            transcriptEvents += 1;
+          }
+        });
+      }
+      if (params.transcriptMtimeMs !== undefined) {
+        advanceTranscriptMutationAtInTransaction(
+          database,
+          params.entry.sessionId,
+          params.transcriptMtimeMs,
+        );
+      } else if (transcriptEvents > 0) {
+        touchTranscriptMutationInTransaction(database, params.entry.sessionId);
+      }
+    }, toDatabaseOptions(resolved));
+    return {
+      sessionId: params.entry.sessionId,
+      sessionKey: resolved.sessionKey,
+      transcriptEvents,
+    };
+  });
+}
+
+/** Appends one raw transcript event to the additive SQLite transcript store. */
+export async function appendSqliteTranscriptEvent(
+  scope: SessionTranscriptAccessScope,
+  event: TranscriptEvent,
+): Promise<void> {
+  assertNonMessageTranscriptEvent(event);
+  const resolved = resolveSqliteTranscriptScope(scope);
+  await runExclusiveSqliteSessionWrite(resolved, async () => {
+    runOpenClawAgentWriteTransaction((database) => {
+      appendTranscriptEventInTransaction(database, resolved, event);
+    }, toDatabaseOptions(resolved));
+  });
+}
+
+/** Appends one raw non-message transcript event synchronously for sync session runtimes. */
+export function appendSqliteTranscriptEventSync(
+  scope: SessionTranscriptAccessScope,
+  event: TranscriptEvent,
+): void {
+  assertNonMessageTranscriptEvent(event);
+  const resolved = resolveSqliteTranscriptScope(scope);
+  runOpenClawAgentWriteTransaction((database) => {
+    const fresh = readSessionEntryRow(database, resolved.sessionKey);
+    if (!fresh || fresh.entry.sessionId !== resolved.sessionId) {
+      return;
+    }
+    appendTranscriptEventInTransaction(database, resolved, event);
+  }, toDatabaseOptions(resolved));
+}
+
+/** Appends a guarded transcript turn and touches its session row in one queued write. */
+export async function appendSqliteExpectedSessionTranscriptTurn(
+  scope: SessionTranscriptWriteScope,
+  options: {
+    config?: import("../types.openclaw.js").OpenClawConfig;
+    cwd?: string;
+    expectedLifecycleRevision?: string;
+    expectedSessionState?: SessionTranscriptTurnExpectedState;
+    expectedSessionId: string;
+    messages: readonly SessionTranscriptTurnMessageAppend[];
+    sessionLifecyclePatch?: SessionTranscriptTurnLifecyclePatch;
+    sessionFile: string;
+    touchSessionEntry?: boolean;
+  },
+): Promise<SqliteExpectedSessionTranscriptTurnResult> {
+  const resolved = resolveSqliteTranscriptScope({
+    ...scope,
+    sessionId: options.expectedSessionId,
+  });
+  return await runExclusiveSqliteSessionWrite(resolved, async () => {
+    const database = openOpenClawAgentDatabase(toDatabaseOptions(resolved));
+    const preparedEntry = readSessionEntryRow(database, resolved.sessionKey);
+    if (!sessionMatchesExpectedTranscriptTurn(preparedEntry, options)) {
+      return sqliteSessionTranscriptTurnRebound(preparedEntry, options.sessionFile);
+    }
+    const messages = await selectAppendableSqliteTranscriptTurnMessages(
+      {
+        agentId: resolved.agentId,
+        sessionFile: options.sessionFile,
+        sessionId: options.expectedSessionId,
+        sessionKey: resolved.sessionKey,
+        ...(scope.storePath ? { storePath: scope.storePath } : {}),
+      },
+      options.messages,
+    );
+    let result: SqliteExpectedSessionTranscriptTurnResult = sqliteSessionTranscriptTurnRebound(
+      preparedEntry,
+      options.sessionFile,
+    );
+    let previousIdentity = new Map<string, SessionEntry>();
+    let currentIdentity = new Map<string, SessionEntry>();
+    runOpenClawAgentWriteTransaction((transactionDb) => {
+      const fresh = readSessionEntryRow(transactionDb, resolved.sessionKey);
+      if (!sessionMatchesExpectedTranscriptTurn(fresh, options)) {
+        result = sqliteSessionTranscriptTurnRebound(fresh, options.sessionFile);
+        return;
+      }
+      const appendedMessages: TranscriptMessageAppendResult<unknown>[] = [];
+      for (const append of messages) {
+        const { shouldAppend: _shouldAppend, ...appendOptions } = append;
+        const appended = appendSqliteTranscriptMessageInTransaction(transactionDb, resolved, {
+          ...appendOptions,
+          ...((append.cwd ?? options.cwd) ? { cwd: append.cwd ?? options.cwd } : {}),
+          ...((append.config ?? options.config) ? { config: append.config ?? options.config } : {}),
+        });
+        if (appended) {
+          appendedMessages.push(appended);
+        }
+      }
+
+      const sessionPatch = buildExpectedTranscriptTurnSessionPatch({
+        appendedMessages,
+        currentEntry: fresh.entry,
+        expectedSessionState: options.expectedSessionState,
+        sessionFile: options.sessionFile,
+        sessionLifecyclePatch: options.sessionLifecyclePatch,
+        touchSessionEntry: options.touchSessionEntry,
+      });
+      const next =
+        Object.keys(sessionPatch).length > 0
+          ? mergeSessionEntry(fresh.entry, sessionPatch)
+          : fresh.entry;
+      if (next !== fresh.entry) {
+        const identityKeys = collectSessionEntryLookupKeys(transactionDb, resolved.sessionKey);
+        previousIdentity = readSqliteSessionIdentitySnapshot(transactionDb, identityKeys);
+        writeSessionEntry(transactionDb, resolved.sessionKey, next);
+        deleteLegacySessionEntryRows(transactionDb, fresh.legacyKeys, resolved.sessionKey);
+        currentIdentity = readSqliteSessionIdentitySnapshot(transactionDb, identityKeys);
+      }
+      result = {
+        appendedMessages,
+        sessionEntry: cloneSessionEntry(next),
+        sessionFile: options.sessionFile,
+      };
+    }, toDatabaseOptions(resolved));
+    emitCommittedSessionIdentityDiff(previousIdentity, currentIdentity);
+    return result;
+  });
+}
+
+function sqliteSessionTranscriptTurnRebound(
+  selected: ResolvedSessionEntryRow | undefined,
+  sessionFile: string,
+): SqliteExpectedSessionTranscriptTurnResult {
+  return {
+    appendedMessages: [],
+    rejectedReason: "session-rebound",
+    sessionEntry: selected?.entry,
+    sessionFile,
+  };
+}
+
+async function selectAppendableSqliteTranscriptTurnMessages(
+  context: SessionTranscriptTurnWriteContext,
+  messages: readonly SessionTranscriptTurnMessageAppend[],
+): Promise<SessionTranscriptTurnMessageAppend[]> {
+  const selected: SessionTranscriptTurnMessageAppend[] = [];
+  for (const append of messages) {
+    const shouldAppend = append.shouldAppend ? await append.shouldAppend(context) : true;
+    if (shouldAppend) {
+      selected.push(append);
+    }
+  }
+  return selected;
+}
+
+/** Appends one transcript message to the additive SQLite transcript store. */
+export async function appendSqliteTranscriptMessage<TMessage>(
+  scope: SessionTranscriptWriteScope,
+  options: TranscriptMessageAppendOptions<TMessage> & {
+    prepareMessageAfterIdempotencyCheck: (message: TMessage) => TMessage | undefined;
+  },
+): Promise<TranscriptMessageAppendResult<TMessage> | undefined>;
+export async function appendSqliteTranscriptMessage<TMessage>(
+  scope: SessionTranscriptWriteScope,
+  options: TranscriptMessageAppendOptions<TMessage>,
+): Promise<TranscriptMessageAppendResult<TMessage>>;
+export async function appendSqliteTranscriptMessage<TMessage>(
+  scope: SessionTranscriptWriteScope,
+  options: TranscriptMessageAppendOptions<TMessage>,
+): Promise<TranscriptMessageAppendResult<TMessage> | undefined> {
+  const resolved = resolveSqliteTranscriptScope(scope);
+  return await runExclusiveSqliteSessionWrite(resolved, async () => {
+    let result: TranscriptMessageAppendResult<TMessage> | undefined;
+    runOpenClawAgentWriteTransaction((database) => {
+      result = appendSqliteTranscriptMessageInTransaction(database, resolved, options);
+    }, toDatabaseOptions(resolved));
+    return result;
+  });
+}
+
+/** Appends one transcript message synchronously for sync session runtimes. */
+export function appendSqliteTranscriptMessageSync<TMessage>(
+  scope: SessionTranscriptWriteScope,
+  options: TranscriptMessageAppendOptions<TMessage>,
+): TranscriptMessageAppendResult<TMessage> | undefined {
+  const resolved = resolveSqliteTranscriptScope(scope);
+  let result: TranscriptMessageAppendResult<TMessage> | undefined;
+  runOpenClawAgentWriteTransaction((database) => {
+    const fresh = readSessionEntryRow(database, resolved.sessionKey);
+    if (!fresh || fresh.entry.sessionId !== resolved.sessionId) {
+      return;
+    }
+    result = appendSqliteTranscriptMessageInTransaction(database, resolved, options);
+  }, toDatabaseOptions(resolved));
+  return result;
+}
+
+/** Runs read/append transcript work under one SQLite writer-queue critical section. */
+export async function withSqliteTranscriptWriteLock<T>(
+  scope: SessionTranscriptWriteScope,
+  run: (context: SqliteTranscriptWriteLockContext) => Promise<T> | T,
+): Promise<T> {
+  const resolved = resolveSqliteTranscriptScope(scope);
+  return await runExclusiveSqliteSessionWrite(resolved, async () => {
+    const database = openOpenClawAgentDatabase(toDatabaseOptions(resolved));
+    let transcriptSnapshot: SqliteTranscriptSnapshotState | undefined;
+    return await run({
+      readEvents: async () => {
+        const snapshot = readSqliteTranscriptSnapshot(database, resolved.sessionId);
+        transcriptSnapshot = { kind: "current", rows: snapshot.rows };
+        return snapshot.events;
+      },
+      replaceEvents: async (events) => {
+        if (transcriptSnapshot?.kind === "stale") {
+          throw new SqliteTranscriptMutationConflictError(resolved.sessionId);
+        }
+        const expectedSnapshot = transcriptSnapshot?.rows;
+        const nextSnapshot = runOpenClawAgentWriteTransaction((writeDatabase) => {
+          if (expectedSnapshot !== undefined) {
+            // The writer queue is process-local. Revalidate after BEGIN IMMEDIATE
+            // so a committed cross-process append cannot be deleted by the rewrite.
+            assertSqliteTranscriptSnapshotUnchanged(
+              writeDatabase,
+              resolved.sessionId,
+              expectedSnapshot,
+            );
+          }
+          replaceSqliteTranscriptEventsInTransaction(writeDatabase, resolved, events);
+          return readSqliteTranscriptSnapshot(writeDatabase, resolved.sessionId).rows;
+        }, toDatabaseOptions(resolved));
+        transcriptSnapshot = { kind: "current", rows: nextSnapshot };
+      },
+      appendMessage: async (options) => {
+        let result: TranscriptMessageAppendResult<unknown> | undefined;
+        const snapshotState = transcriptSnapshot;
+        let nextSnapshotState = snapshotState;
+        runOpenClawAgentWriteTransaction((writeDatabase) => {
+          const snapshotStillCurrent =
+            snapshotState?.kind === "current"
+              ? isSqliteTranscriptSnapshotUnchanged(
+                  writeDatabase,
+                  resolved.sessionId,
+                  snapshotState.rows,
+                )
+              : false;
+          result = appendSqliteTranscriptMessageInTransaction(writeDatabase, resolved, options);
+          if (snapshotState?.kind === "current") {
+            nextSnapshotState = snapshotStillCurrent
+              ? {
+                  kind: "current",
+                  rows: readSqliteTranscriptSnapshot(writeDatabase, resolved.sessionId).rows,
+                }
+              : { kind: "stale" };
+          }
+        }, toDatabaseOptions(resolved));
+        transcriptSnapshot = nextSnapshotState;
+        return result as TranscriptMessageAppendResult<typeof options.message> | undefined;
+      },
+    });
+  });
+}
+
+/** Runs synchronous transcript work under one writer queue and SQLite transaction. */
+export async function withSqliteTranscriptWriteTransaction<T>(
+  scope: SessionTranscriptWriteScope,
+  run: (context: { sessionFile: string }) => T,
+): Promise<T> {
+  const resolved = resolveSqliteTranscriptScope(scope);
+  return await runExclusiveSqliteSessionWrite(resolved, async () =>
+    runOpenClawAgentWriteTransaction(
+      () => run({ sessionFile: formatSqliteSessionMarkerForScope(resolved) }),
+      toDatabaseOptions(resolved),
+      { operationLabel: "session.transcript.batch" },
+    ),
+  );
+}
+
+function isSqliteTranscriptSnapshotUnchanged(
+  database: OpenClawAgentDatabase,
+  sessionId: string,
+  expected: readonly SqliteTranscriptSnapshotRow[],
+): boolean {
+  const current = readSqliteTranscriptSnapshot(database, sessionId).rows;
+  return (
+    current.length === expected.length &&
+    current.every(
+      (row, index) =>
+        row.seq === expected[index]?.seq && row.eventJson === expected[index]?.eventJson,
+    )
+  );
+}
+
+function assertSqliteTranscriptSnapshotUnchanged(
+  database: OpenClawAgentDatabase,
+  sessionId: string,
+  expected: readonly SqliteTranscriptSnapshotRow[],
+): void {
+  if (!isSqliteTranscriptSnapshotUnchanged(database, sessionId, expected)) {
+    throw new SqliteTranscriptMutationConflictError(sessionId);
+  }
+}
+
+function appendSqliteTranscriptMessageInTransaction<TMessage>(
+  database: OpenClawAgentDatabase,
+  resolved: ResolvedTranscriptScope,
+  options: TranscriptMessageAppendOptions<TMessage>,
+): TranscriptMessageAppendResult<TMessage> | undefined {
+  const idempotencyKey = readMessageIdempotencyKey(options.message);
+  if (idempotencyKey && options.idempotencyLookup !== "caller-checked") {
+    const existing = readTranscriptMessageByScopedIdempotencyKey(
+      database,
+      resolved,
+      idempotencyKey,
+      options.idempotencyLookup,
+    );
+    if (existing) {
+      return {
+        appended: false,
+        message: existing.message as TMessage,
+        messageId: existing.messageId,
+      };
+    }
+  }
+
+  const prepared = options.prepareMessageAfterIdempotencyCheck
+    ? options.prepareMessageAfterIdempotencyCheck(options.message)
+    : options.message;
+  if (prepared === undefined) {
+    return undefined;
+  }
+
+  const messageId = options.eventId ?? randomUUID();
+  const now = options.now ?? Date.now();
+  const finalMessage = redactTranscriptMessageForStorage(prepared, options);
+  ensureTranscriptHeader(database, resolved, options.cwd, now);
+  const parentId =
+    options.parentId === undefined
+      ? readActiveTranscriptAppendParentId(database, resolved.sessionId)
+      : options.parentId;
+  const event = {
+    type: "message",
+    id: messageId,
+    parentId: parentId ?? null,
+    timestamp: resolveTimestampMsToIsoString(now),
+    message: finalMessage,
+  };
+  const appended = appendTranscriptEventInTransaction(database, resolved, event, {
+    dedupeByMessageIdempotency:
+      options.idempotencyLookup !== "caller-checked" &&
+      options.idempotencyLookup !== "scan-assistant",
+  });
+  if (!appended && idempotencyKey && options.idempotencyLookup !== "caller-checked") {
+    const existing = readTranscriptMessageByScopedIdempotencyKey(
+      database,
+      resolved,
+      idempotencyKey,
+      options.idempotencyLookup,
+    );
+    if (existing) {
+      return {
+        appended: false,
+        message: existing.message as TMessage,
+        messageId: existing.messageId,
+      };
+    }
+  }
+  if (!appended) {
+    const existing = readTranscriptMessageByEventId(database, resolved, messageId);
+    if (existing) {
+      return {
+        appended: false,
+        message: existing.message as TMessage,
+        messageId: existing.messageId,
+      };
+    }
+  }
+  if (!appended) {
+    throw new Error(`SQLite transcript append did not insert message ${messageId}.`);
+  }
+  return {
+    appended: true,
+    message: finalMessage,
+    messageId,
+  };
+}
+
+function assertNonMessageTranscriptEvent(event: TranscriptEvent): void {
+  if (!event || typeof event !== "object" || Array.isArray(event)) {
+    return;
+  }
+  // Message records require parent-link, idempotency, and redaction handling
+  // from appendSqliteTranscriptMessage; raw event writes would bypass those invariants.
+  if ((event as { type?: unknown }).type === "message") {
+    throw new Error(
+      "appendSqliteTranscriptEvent cannot write message transcript records; use appendSqliteTranscriptMessage instead.",
+    );
+  }
+}
